refactor(plugin-support, plugin-observability): move feedback UI to plugin-support
plugin-observability becomes a pure telemetry plugin: it still contributes the
Observability and LogDownloader capabilities, but no longer owns any
user-facing UI. plugin-support now owns:

- FeedbackForm component (with story).
- FeedbackPanel container that consumes Observability + LogDownloader and
  shows a toast on submit.
- CaptureUserFeedback operation that forwards to the Observability capability.
- The deck `--help` companion surface previously contributed by observability.
- The corresponding translation strings.

UserFeedback schema annotations are inlined as English strings (to break a
types -> translations import cycle); user-facing strings still come from
translations at render time.

composer-app/ResetDialog imports FeedbackForm and the UserFeedback type from
plugin-support instead of plugin-observability.
