+++ packages/apps/composer-app/src/components/ResetDialog/ResetDialog.tsx
@@ -6,8 +6,8 @@ import React, { useCallback, useEffect, useState } from 'react';
 
 import { type IdbLogStore } from '@dxos/log-store-idb';
 import { type Observability } from '@dxos/observability';
-import { type ObservabilityOperation } from '@dxos/plugin-observability';
-import { FeedbackForm } from '@dxos/plugin-observability/components';
+import { type SupportOperation } from '@dxos/plugin-support';
+import { FeedbackForm } from '@dxos/plugin-support/components';
 import {
   AlertDialog,
   type AlertDialogRootProps,
@@ -93,7 +93,7 @@ export const ResetDialog = ({
   }, [download, logStore]);
 
   const handleSaveFeedback = useCallback(
-    async (values: ObservabilityOperation.UserFeedback) => {
+    async (values: SupportOperation.UserFeedback) => {
       if (!observabilityProp) {
         return;
       }
