+++ packages/plugins/plugin-observability/moon.yml
@@ -14,7 +14,6 @@ tasks:
       - '--entryPoint=src/ObservabilityPlugin.node.ts'
       - '--entryPoint=src/capabilities/index.ts'
       - '--entryPoint=src/components/index.ts'
-      - '--entryPoint=src/containers/index.ts'
       - '--entryPoint=src/meta.ts'
       - '--entryPoint=src/operations/index.ts'
       - '--entryPoint=src/plugin.ts'
