+++ packages/plugins/plugin-observability/package.json
@@ -23,11 +23,6 @@
       "types": "./dist/types/src/components/index.d.ts",
       "default": "./dist/lib/neutral/components/index.mjs"
     },
-    "#containers": {
-      "source": "./src/containers/index.ts",
-      "types": "./dist/types/src/containers/index.d.ts",
-      "default": "./dist/lib/neutral/containers/index.mjs"
-    },
     "#meta": {
       "source": "./src/meta.ts",
       "types": "./dist/types/src/meta.d.ts",
@@ -63,11 +58,6 @@
       "source": "./src/index.ts",
       "types": "./dist/types/src/index.d.ts",
       "default": "./dist/lib/neutral/index.mjs"
-    },
-    "./components": {
-      "source": "./src/components/index.ts",
-      "types": "./dist/types/src/components/index.d.ts",
-      "default": "./dist/lib/neutral/components/index.mjs"
     },
     "./plugin": {
       "source": "./src/plugin.ts",
