+++ packages/plugins/plugin-observability/src/components/index.ts
@@ -4,6 +4,4 @@
 
 import { type ComponentType, lazy } from 'react';
 
-export * from './FeedbackForm';
-
 export const ObservabilitySettings: ComponentType<any> = lazy(() => import('./ObservabilitySettings'));
