--- packages/plugins/plugin-observability/src/containers/HelpContainer/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2025 DXOS.org
-//
-
-export { HelpContainer as default } from './HelpContainer';
