--- packages/plugins/plugin-observability/src/containers/index.ts
@@ -1,7 +1,0 @@
-//
-// Copyright 2025 DXOS.org
-//
-
-import { type ComponentType, lazy } from 'react';
-
-export const HelpContainer: ComponentType<any> = lazy(() => import('./HelpContainer'));
