+++ packages/plugins/plugin-observability/src/operations/index.ts
@@ -5,7 +5,6 @@
 import { OperationHandlerSet } from '@dxos/compute';
 
 export const ObservabilityOperationHandlerSet = OperationHandlerSet.lazy(
-  () => import('./capture-user-feedback'),
   () => import('./send-event'),
   () => import('./toggle'),
 );
