+++ packages/plugins/plugin-observability/src/translations.ts
@@ -21,14 +21,6 @@ export const translations = [
         // TODO: Add link about telemetry privacy. Make it clearer that user data is not collected.
         'observability.description':
           'When enabled, basic usage data will used to improve the product. This may include performance metrics, error logs, and usage data. No personally identifiable information, other than your public key, is included with this data and no private data ever leaves your devices.',
-        'help.label': 'Feedback & Support',
-        'feedback-textarea.label': 'Feedback',
-        'feedback-textarea.placeholder': 'Please enter your feedback, technical issue, or feature request.',
-        'include-debug-logs.label': 'Include debug logs',
-        'send-feedback.label': 'Send Feedback',
-        'download-logs.label': 'Download logs',
-        'feedback-toast.label': 'Thank you for your feedback!',
-        'feedback-toast.description': 'We will review your feedback and get back to you as soon as possible.',
       },
     },
   },
