+++ packages/plugins/plugin-observability/src/types/ObservabilityOperation.ts
@@ -8,31 +8,8 @@ import * as Schema from 'effect/Schema';
 
 import { Capability } from '@dxos/app-framework';
 import { Operation } from '@dxos/compute';
-import { Format } from '@dxos/echo';
 
 import { meta } from '#meta';
-import { translations } from '#translations';
-
-const t = translations[0]['en-US'][meta.id];
-
-export const UserFeedback = Schema.Struct({
-  message: Format.Text.pipe(
-    Schema.nonEmptyString(),
-    Schema.maxLength(4_096),
-    Schema.annotations({
-      title: t['feedback-textarea.label'],
-      description: t['feedback-textarea.placeholder'],
-    }),
-  ),
-  includeLogs: Schema.Boolean.pipe(
-    Schema.annotations({
-      title: t['include-debug-logs.label'],
-    }),
-    Schema.optional,
-  ),
-});
-
-export type UserFeedback = Schema.Schema.Type<typeof UserFeedback>;
 
 const OBSERVABILITY_OPERATION = `${meta.id}.operation`;
 
@@ -47,17 +24,6 @@ export const Toggle = Operation.make({
     state: Schema.optional(Schema.Boolean),
   }),
   output: Schema.Boolean,
-});
-
-export const CaptureUserFeedback = Operation.make({
-  meta: {
-    key: `${OBSERVABILITY_OPERATION}.capture-feedback`,
-    name: 'Capture User Feedback',
-    description: 'Capture user feedback.',
-  },
-  services: [Capability.Service],
-  input: UserFeedback,
-  output: Schema.Void,
 });
 
 export const SendEvent = Operation.make({
