+++ packages/plugins/plugin-support/moon.yml
@@ -11,6 +11,7 @@ tasks:
       - '--entryPoint=src/SupportPlugin.tsx'
       - '--entryPoint=src/blueprints/index.ts'
       - '--entryPoint=src/capabilities/index.ts'
+      - '--entryPoint=src/components/index.ts'
       - '--entryPoint=src/containers/index.ts'
       - '--entryPoint=src/meta.ts'
       - '--entryPoint=src/operations/index.ts'
