+++ packages/plugins/plugin-support/src/components/FeedbackForm/index.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2024 DXOS.org
+// Copyright 2026 DXOS.org
 //
 
 export * from './FeedbackForm';
