+++ packages/plugins/plugin-support/src/components/index.ts
@@ -2,4 +2,4 @@
 // Copyright 2026 DXOS.org
 //
 
-export {};
+export * from './FeedbackForm';
