+++ packages/plugins/plugin-support/src/containers/FeedbackPanel/FeedbackPanel.tsx
@@ -1,5 +1,5 @@
 //
-// Copyright 2025 DXOS.org
+// Copyright 2026 DXOS.org
 //
 
 import * as Effect from 'effect/Effect';
@@ -8,16 +8,16 @@ import React, { useCallback, useState } from 'react';
 import { useCapabilities, useCapability, useOperationInvoker } from '@dxos/app-framework/ui';
 import { LayoutOperation } from '@dxos/app-toolkit';
 import { runAndForwardErrors } from '@dxos/effect';
+import { ObservabilityCapabilities } from '@dxos/plugin-observability';
 import { useAsyncEffect } from '@dxos/react-hooks';
 import { osTranslations } from '@dxos/ui-theme';
 
 import { FeedbackForm } from '#components';
 import { meta } from '#meta';
-import { ObservabilityOperation } from '#types';
-import { ObservabilityCapabilities } from '#types';
+import { SupportOperation } from '#types';
 
 /** Renders the feedback form, disabling it when the feedback survey is unavailable. */
-export const HelpContainer = () => {
+export const FeedbackPanel = () => {
   const { invokePromise } = useOperationInvoker();
   const observability = useCapability(ObservabilityCapabilities.Observability);
   const [downloadLogs] = useCapabilities(ObservabilityCapabilities.LogDownloader);
@@ -38,8 +38,8 @@ export const HelpContainer = () => {
   );
 
   const handleSave = useCallback(
-    async (values: ObservabilityOperation.UserFeedback) => {
-      await invokePromise(ObservabilityOperation.CaptureUserFeedback, values);
+    async (values: SupportOperation.UserFeedback) => {
+      await invokePromise(SupportOperation.CaptureUserFeedback, values);
       await invokePromise(LayoutOperation.UpdateComplementary, { state: 'collapsed' });
       await invokePromise(LayoutOperation.AddToast, {
         id: `${meta.id}.feedback-success`,
