+++ packages/plugins/plugin-support/src/containers/FeedbackPanel/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export { FeedbackPanel as default } from './FeedbackPanel';
