+++ packages/plugins/plugin-support/src/containers/index.ts
@@ -4,4 +4,5 @@
 
 import { type ComponentType, lazy } from 'react';
 
+export const FeedbackPanel: ComponentType<any> = lazy(() => import('./FeedbackPanel'));
 export const SupportArticle: ComponentType<any> = lazy(() => import('./SupportArticle'));
