+++ packages/plugins/plugin-support/src/operations/capture-feedback.ts
@@ -1,16 +1,17 @@
 //
-// Copyright 2025 DXOS.org
+// Copyright 2026 DXOS.org
 //
 
 import * as Effect from 'effect/Effect';
 
 import { Capability } from '@dxos/app-framework';
 import { Operation } from '@dxos/compute';
+import { ObservabilityCapabilities } from '@dxos/plugin-observability';
 
-import { ObservabilityCapabilities, ObservabilityOperation } from '../types';
+import { SupportOperation } from '../types';
 
-const handler: Operation.WithHandler<typeof ObservabilityOperation.CaptureUserFeedback> =
-  ObservabilityOperation.CaptureUserFeedback.pipe(
+const handler: Operation.WithHandler<typeof SupportOperation.CaptureUserFeedback> =
+  SupportOperation.CaptureUserFeedback.pipe(
     Operation.withHandler(
       Effect.fnUntraced(function* (input) {
         const observability = yield* Capability.get(ObservabilityCapabilities.Observability);
