+++ packages/plugins/plugin-support/src/translations.ts
@@ -32,6 +32,15 @@ export const translations = [
         'mark-in-progress.button': 'Mark in progress',
         'resolve.button': 'Resolve',
         'reopen.button': 'Reopen',
+        // Feedback (deck-companion--help surface).
+        'help.label': 'Feedback & Support',
+        'feedback-textarea.label': 'Feedback',
+        'feedback-textarea.placeholder': 'Please enter your feedback, technical issue, or feature request.',
+        'include-debug-logs.label': 'Include debug logs',
+        'send-feedback.label': 'Send Feedback',
+        'download-logs.label': 'Download logs',
+        'feedback-toast.label': 'Thank you for your feedback!',
+        'feedback-toast.description': 'We will review your feedback and get back to you as soon as possible.',
       },
     },
   },
